chore: Extract test notification into shared utility
- Move test notification logic from settings into dedicated service
- Remove test button from settings header
